[refactoring] import 호출 최적화
