+++ src/main/java/org/example/servlet/view/ArticleDetailView.java
@@ -6,14 +6,12 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.example.config.DataHandler;
+import java.io.IOException;
+import org.example.constance.DataHandler;
 import org.example.data.ArticleDataHandler;
-import org.example.data.ArticleDataHandlerMySql;
 import org.example.domain.Article;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
 
 @WebServlet(urlPatterns = {"/articles/*"})
 public class ArticleDetailView extends HttpServlet {
